+++ plugins/single_plugins/wsets.cpp
@@ -23,6 +23,7 @@
 #include <wayfire/config/types.hpp>
 #include <wayfire/output-layout.hpp>
 #include <wayfire/bindings-repository.hpp>
+#include <list>
 
 
 class wayfire_wsets_plugin_t : public wf::plugin_interface_t
