+++ plugins/wm-actions/wm-actions.cpp
@@ -18,6 +18,7 @@
 #include "wayfire/seat.hpp"
 #include "wayfire/nonstd/reverse.hpp"
 #include "wm-actions-signals.hpp"
+#include "wayfire/nonstd/reverse.hpp"
 
 class always_on_top_root_node_t : public wf::scene::output_node_t
 {
