+++ plugins/wobbly/wobbly.cpp
@@ -932,6 +932,12 @@ class wayfire_wobbly : public wf::plugin_interface_t
     void adjust_wobbly(wobbly_signal *data)
     {
         auto tr_manager = data->view->get_transformed_node();
+
+        if ((data->events == WOBBLY_EVENT_ACTIVATE) && tr_manager->get_transformer("wobbly"))
+        {
+            return;
+        }
+
         if ((data->events & (WOBBLY_EVENT_GRAB | WOBBLY_EVENT_ACTIVATE)) &&
             !tr_manager->get_transformer<wobbly_transformer_node_t>("wobbly"))
         {
