+++ src/api/wayfire/signal-provider.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <functional>
-#include <memory>
 #include <unordered_set>
 #include <unordered_map>
 #include <wayfire/nonstd/safe-list.hpp>
